+++ web/server/current_packages.txt
@@ -1,0 +1,260 @@
+absl-py==2.1.0
+accelerate==1.0.1
+aiohappyeyeballs==2.4.3
+aiohttp==3.10.10
+aioice==0.9.0
+aiortc==1.9.0
+aiosignal==1.3.1
+altair==5.4.1
+annotated-types==0.7.0
+anyio==4.4.0
+appnope==0.1.4
+argon2-cffi==23.1.0
+argon2-cffi-bindings==21.2.0
+arrow==1.3.0
+asgiref==3.8.1
+asttokens==2.4.1
+astunparse==1.6.3
+async-lru==2.0.4
+async-timeout==4.0.3
+asyncio==3.4.3
+attrs==24.2.0
+av==12.3.0
+babel==2.16.0
+beautifulsoup4==4.12.3
+bleach==6.1.0
+blinker==1.8.2
+blis==1.0.1
+cachetools==5.5.0
+catalogue==2.0.10
+certifi==2024.8.30
+cffi==1.17.1
+charset-normalizer==3.3.2
+click==8.1.7
+cloudpathlib==0.20.0
+comm==0.2.2
+confection==0.1.5
+contourpy==1.3.0
+cryptography==43.0.3
+cycler==0.12.1
+cymem==2.0.8
+datasets==3.0.1
+debugpy==1.8.5
+decorator==5.1.1
+defusedxml==0.7.1
+dill==0.3.8
+Django==4.2.13
+django-cors-headers==3.13.0
+django-extensions==3.2.1
+djangorestframework==3.14.0
+dnspython==2.6.1
+en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
+exceptiongroup==1.2.2
+executing==2.1.0
+Faker==30.8.0
+fastjsonschema==2.20.0
+filelock==3.16.1
+Flask==3.0.3
+flask-ngrok==0.0.25
+flatbuffers==1.12
+fonttools==4.53.1
+fqdn==1.5.1
+frozenlist==1.4.1
+fsspec==2024.6.1
+gast==0.4.0
+gitdb==4.0.11
+GitPython==3.1.43
+google-auth==2.34.0
+google-auth-oauthlib==0.4.6
+google-crc32c==1.6.0
+google-pasta==0.2.0
+grpcio==1.66.1
+gTTS==2.5.3
+h11==0.14.0
+h5py==3.11.0
+html5lib-modern==1.2
+httpcore==1.0.5
+httpx==0.27.2
+huggingface==0.0.1
+huggingface-hub==0.26.1
+idna==3.8
+ifaddr==0.2.0
+importlib_metadata==8.4.0
+importlib_resources==6.4.4
+ipykernel==6.29.5
+ipython==8.18.1
+ipywidgets==8.1.5
+isodate==0.7.2
+isoduration==20.11.0
+itsdangerous==2.2.0
+jax==0.4.30
+jaxlib==0.4.30
+jedi==0.19.1
+Jinja2==3.1.4
+joblib==1.4.2
+json5==0.9.25
+jsonpointer==3.0.0
+jsonschema==4.23.0
+jsonschema-specifications==2023.12.1
+jupyter==1.1.1
+jupyter-console==6.6.3
+jupyter-events==0.10.0
+jupyter-lsp==2.2.5
+jupyter_client==8.6.2
+jupyter_core==5.7.2
+jupyter_server==2.14.2
+jupyter_server_terminals==0.5.3
+jupyterlab==4.2.5
+jupyterlab_pygments==0.3.0
+jupyterlab_server==2.27.3
+jupyterlab_widgets==3.0.13
+keras==2.9.0
+Keras-Preprocessing==1.1.2
+keras-tuner==1.1.3
+kiwisolver==1.4.7
+kt-legacy==1.0.5
+langcodes==3.4.1
+language_data==1.2.0
+libclang==18.1.1
+lora==0.3.0
+marisa-trie==1.2.1
+Markdown==3.7
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.6.0
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+mediapipe==0.9.3.0
+mediapipe-silicon==0.8.10.1
+mistune==3.0.2
+ml-dtypes==0.4.0
+mpmath==1.3.0
+multidict==6.1.0
+multiprocess==0.70.16
+murmurhash==1.0.10
+namex==0.0.8
+narwhals==1.10.0
+nbclient==0.10.0
+nbconvert==7.16.4
+nbformat==5.10.4
+nest-asyncio==1.6.0
+networkx==3.2.1
+notebook==7.2.2
+notebook_shim==0.2.4
+numpy==1.23.3
+oauthlib==3.2.2
+openai==0.28.0
+opencv-contrib-python==4.10.0.84
+opencv-python==4.6.0.66
+opt-einsum==3.3.0
+optree==0.12.1
+overrides==7.7.0
+packaging==24.1
+pandas==1.4.3
+pandocfilters==1.5.1
+parso==0.8.4
+pexpect==4.9.0
+pillow==10.4.0
+platformdirs==4.3.2
+playsound==1.3.0
+plotly==5.24.1
+preshed==3.0.9
+prometheus_client==0.20.0
+prompt_toolkit==3.0.47
+propcache==0.2.0
+protobuf==3.20.3
+psutil==6.0.0
+ptyprocess==0.7.0
+pure_eval==0.2.3
+pyarrow==17.0.0
+pyasn1==0.6.0
+pyasn1_modules==0.4.0
+pycparser==2.22
+pydantic==2.9.2
+pydantic_core==2.23.4
+pydeck==0.9.1
+pydub==0.25.1
+pyee==12.0.0
+Pygments==2.18.0
+pylibsrtp==0.10.0
+pymongo==4.8.0
+pyOpenSSL==24.2.1
+pyparsing==3.1.4
+python-dateutil==2.9.0.post0
+python-json-logger==2.0.7
+pytz==2024.1
+PyYAML==6.0.2
+pyzmq==26.2.0
+rdflib==7.1.0
+referencing==0.35.1
+regex==2024.9.11
+requests==2.32.3
+requests-oauthlib==2.0.0
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+rich==13.8.1
+rpds-py==0.20.0
+rsa==4.9
+safetensors==0.4.5
+scikit-learn==1.1.2
+scipy==1.13.1
+seaborn==0.12.0
+Send2Trash==1.8.3
+shellingham==1.5.4
+six==1.16.0
+smart-open==7.0.5
+smmap==5.0.1
+sniffio==1.3.1
+sounddevice==0.5.0
+soupsieve==2.6
+spacy==3.8.2
+spacy-legacy==3.0.12
+spacy-loggers==1.0.5
+sqlparse==0.5.1
+srsly==2.4.8
+stack-data==0.6.3
+streamlit==1.38.0
+streamlit-drawable-canvas==0.9.3
+streamlit-webrtc==0.47.9
+sympy==1.13.1
+tenacity==8.5.0
+tensorboard==2.9.1
+tensorboard-data-server==0.6.1
+tensorboard-plugin-wit==1.8.1
+tensorflow==2.17.0
+tensorflow-estimator==2.9.0
+tensorflow-io-gcs-filesystem==0.37.1
+tensorflow-macos==2.9.2
+termcolor==2.4.0
+terminado==0.18.1
+thinc==8.3.2
+threadpoolctl==3.5.0
+tinycss2==1.3.0
+tokenizers==0.20.1
+toml==0.10.2
+tomli==2.0.1
+torch==2.5.0
+tornado==6.4.1
+tqdm==4.66.5
+traitlets==5.14.3
+transformers==4.45.2
+typer==0.12.5
+types-python-dateutil==2.9.0.20240906
+typing_extensions==4.12.2
+tzdata==2024.1
+uri-template==1.3.0
+urllib3==2.2.2
+wasabi==1.1.3
+watchdog==5.0.3
+wcwidth==0.2.13
+weasel==0.4.1
+webcolors==24.8.0
+webencodings==0.5.1
+websocket-client==1.8.0
+Werkzeug==3.0.4
+widgetsnbextension==4.0.13
+wrapt==1.16.0
+xxhash==3.5.0
+yarl==1.16.0
+yellowbrick==1.5
+zipp==3.20.1
